requirements.txt
Django
djangorestframework
coreapi
pygments
markdown
